+++ src/components/battle/select/hooks/useVoteResult.ts
